Move LocaleContext file
